+++ src/version-diff.ads
@@ -18,6 +18,7 @@ package Version.Diff is
       Shortstat      : Boolean := False;
       Raw            : Boolean := False;
       Compact_Summary : Boolean := False;
+      Stat_Width      : Natural := 0;
       Detect_Renames : Rename_Detection := Renames_Default;
       Rename_Score   : Natural := 0;
       Rename_Limit   : Natural := 0;